Rename doc2 -> doc
