+++ lapce-app/src/config/editor.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use structdesc::FieldNames;
 
-use crate::doc2::RenderWhitespace;
+use crate::doc::RenderWhitespace;
 
 pub const SCALE_OR_SIZE_LIMIT: f64 = 5.0;
