+++ lapce-app/src/inline_completion.rs
@@ -11,7 +11,7 @@ use lapce_core::{
 
 use lsp_types::InsertTextFormat;
 
-use crate::{config::LapceConfig, doc2::Doc, editor::EditorData, snippet::Snippet};
+use crate::{config::LapceConfig, doc::Doc, editor::EditorData, snippet::Snippet};
 
 // TODO: we could integrate completion lens with this, so it is considered at the same time
